dependencies/requirements.txt
pillow
tkinterdnd2
pandas
requests